add ability to support debug
