+++ pkg/microservice/aslan/core/common/service/workflowcontroller/workflow.go
@@ -18,8 +18,14 @@ package workflowcontroller
 
 import (
 	"context"
+	"encoding/json"
 	"errors"
 	"fmt"
+	e "github.com/koderover/zadig/v2/pkg/tool/errors"
+	"github.com/koderover/zadig/v2/pkg/tool/kube/getter"
+	"github.com/koderover/zadig/v2/pkg/tool/kube/podexec"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/labels"
 	"strings"
 	"time"
 
@@ -44,6 +50,26 @@ import (
 	"github.com/koderover/zadig/v2/pkg/tool/log"
 )
 
+const (
+	checkShellStepStart  = "ls /zadig/debug/shell_step"
+	checkShellStepDone   = "ls /zadig/debug/shell_step_done"
+	setOrUnsetBreakpoint = "%s /zadig/debug/breakpoint_%s"
+)
+
+const (
+	WorkflowDebugEventEnableDebug   = "EnableDebug"
+	WorkflowDebugEventDeleteDebug   = "DeleteDebug"
+	WorkflowDebugEventSetBreakPoint = "SetBreakpoint"
+)
+
+type WorkflowDebugEvent struct {
+	EventType string `json:"event_type"`
+	JobName   string `json:"job_name"`
+	TaskID    int64  `json:"task_id"`
+	Set       bool   `json:"set"`
+	Position  string `json:"position"`
+}
+
 type workflowCtl struct {
 	workflowTask       *commonmodels.WorkflowTask
 	globalContextMutex *cache.RedisLock
@@ -56,8 +82,7 @@ func NewWorkflowController(workflowTask *commonmodels.WorkflowTask, logger *zap.
 	ctl := &workflowCtl{
 		workflowTask: workflowTask,
 		logger:       logger,
-		prefix:       fmt.Sprintf("workflowctl-%s-%d", workflowTask.WorkflowName, workflowTask.TaskID),
-	}
+		prefix:       fmt.Sprintf("workflowctl-%s-%d", workflowTask.WorkflowName, workflowTask.TaskID)}
 	ctl.globalContextMutex = cache.NewRedisLock(fmt.Sprintf("%s-global-context", ctl.prefix))
 	ctl.ack = ctl.updateWorkflowTask
 	return ctl
@@ -111,10 +136,38 @@ func (c *workflowCtl) setWorkflowStatus(status config.Status) {
 	c.ack()
 }
 
+func WorkflowDebugLockKey(workflowName string, taskID int64) string {
+	return fmt.Sprintf("workflowctl-debug-lock-%s-%d", workflowName, taskID)
+}
+
+func WorkflowDebugChanKey(workflowName string, taskID int64) string {
+	return fmt.Sprintf("workflowctl-debug-lock-%s-%d", workflowName, taskID)
+}
+
+func (c *workflowCtl) handleDebugEvent(bytes string) error {
+	event := &WorkflowDebugEvent{}
+	err := json.Unmarshal([]byte(bytes), event)
+	if err != nil {
+		return fmt.Errorf("failed to unmarshal event, err: %s", err)
+	}
+	log.Infof("handing workflow debug event: %s/%d, event: %s", c.workflowTask.WorkflowName, c.workflowTask.TaskID, event.EventType)
+
+	switch event.EventType {
+	case WorkflowDebugEventSetBreakPoint:
+		err = c.handleWorkflowBreakpoint(event.JobName, event.Position, event.Set)
+	case WorkflowDebugEventEnableDebug:
+		err = c.enableWorkflowDebug()
+	case WorkflowDebugEventDeleteDebug:
+		err = c.stopWorkflowDebug(event.JobName, event.Position)
+	default:
+		err = fmt.Errorf("unknow debug event: %s", err)
+	}
+	return err
+}
+
 func (c *workflowCtl) Run(ctx context.Context, concurrency int) {
 
-	// TODO optimize me, this doesn't work when aslan running in multiple replicas
-	addWorkflowTaskInMap(c.workflowTask.WorkflowName, c.workflowTask.TaskID, c.workflowTask, c.ack)
+	addWorkflowTaskInMap(c.workflowTask.WorkflowName, c.workflowTask.TaskID, c.workflowTask)
 	defer removeWorkflowTaskInMap(c.workflowTask.WorkflowName, c.workflowTask.TaskID)
 
 	c.workflowTask.Status = config.StatusRunning
@@ -133,15 +186,28 @@ func (c *workflowCtl) Run(ctx context.Context, concurrency int) {
 
 	// sub cancel signal from redis
 	cancelChan, closeFunc := cache.NewRedisCache(config2.RedisCommonCacheTokenDB()).Subscribe(fmt.Sprintf("workflowctl-cancel-%s-%d", c.workflowTask.WorkflowName, c.workflowTask.TaskID))
-	defer func() { _ = closeFunc() }()
+	debugChan, closeDebugChanFunc := cache.NewRedisCache(config2.RedisCommonCacheTokenDB()).Subscribe(WorkflowDebugChanKey(c.workflowTask.WorkflowName, c.workflowTask.TaskID))
+	defer func() {
+		log.Infof("-------- closing pubsub channels")
+		_ = closeFunc()
+		_ = closeDebugChanFunc()
+	}()
 
 	// receiving cancel signal from redis
 	go func() {
-		select {
-		case <-cancelChan:
-			cancel()
-		case <-ctx.Done():
-			return
+		for {
+			select {
+			case data := <-debugChan:
+				err := c.handleDebugEvent(data.Payload)
+				if err != nil {
+					c.logger.Errorf(fmt.Sprintf("workflow ctl run err: %s", err))
+				}
+			case <-cancelChan:
+				cancel()
+				return
+			case <-ctx.Done():
+				return
+			}
 		}
 	}()
 
@@ -176,6 +242,246 @@ func (c *workflowCtl) Run(ctx context.Context, concurrency int) {
 	RunStages(ctx, c.workflowTask.Stages, workflowCtx, concurrency, c.logger, c.ack)
 	updateworkflowStatus(c.workflowTask)
 	c.workflowTask.GlobalContext = c.getGlobalContextAll()
+}
+
+func (c *workflowCtl) handleWorkflowBreakpoint(jobName, position string, set bool) error {
+	workflowDebugLock := cache.NewRedisLockWithExpiry(WorkflowDebugLockKey(c.workflowTask.WorkflowName, c.workflowTask.TaskID), time.Second*5)
+	err := workflowDebugLock.Lock()
+	if err != nil {
+		return e.ErrStopDebugShell.AddDesc(fmt.Sprintf("failed to acquire lock when setting breakpoint, err: %s", err))
+	}
+	defer workflowDebugLock.Unlock()
+
+	logger := c.logger
+	var ack func()
+	defer func() {
+		if ack != nil {
+			c.ack()
+		}
+	}()
+	var task *commonmodels.JobTask
+FOR:
+	for _, stage := range c.workflowTask.Stages {
+		for _, jobTask := range stage.Jobs {
+			if jobTask.Name == jobName {
+				task = jobTask
+				break FOR
+			}
+		}
+	}
+	if task == nil {
+		logger.Error("set workflowTaskV4 breakpoint failed: not found job")
+		return e.ErrSetBreakpoint.AddDesc("当前任务不存在")
+	}
+	// job task has not run, update data in memory and ack
+	if task.Status == "" {
+		switch position {
+		case "before":
+			task.BreakpointBefore = set
+			ack = c.ack
+		case "after":
+			task.BreakpointAfter = set
+			ack = c.ack
+		}
+		logger.Infof("set workflowTaskV4 breakpoint success: %s-%s %v", jobName, position, set)
+		return nil
+	}
+
+	jobTaskSpec := &commonmodels.JobTaskFreestyleSpec{}
+	if err := commonmodels.IToi(task.Spec, jobTaskSpec); err != nil {
+		logger.Errorf("set workflowTaskV4 breakpoint failed: IToi %v", err)
+		return e.ErrSetBreakpoint.AddDesc("修改断点意外失败: convert job task spec")
+	}
+
+	kubeClient, err := kubeclient.GetKubeClient(config.HubServerAddress(), jobTaskSpec.Properties.ClusterID)
+	if err != nil {
+		log.Errorf("set workflowTaskV4 breakpoint failed: get kube client error: %s", err)
+		return e.ErrSetBreakpoint.AddDesc("修改断点意外失败: get kube client")
+	}
+	clientSet, err := kubeclient.GetClientset(config.HubServerAddress(), jobTaskSpec.Properties.ClusterID)
+	if err != nil {
+		log.Errorf("set workflowTaskV4 breakpoint failed: get kube client set error: %s", err)
+		return e.ErrSetBreakpoint.AddDesc("修改断点意外失败: get kube client set")
+	}
+	restConfig, err := kubeclient.GetRESTConfig(config.HubServerAddress(), jobTaskSpec.Properties.ClusterID)
+	if err != nil {
+		log.Errorf("set workflowTaskV4 breakpoint failed: get kube rest config error: %s", err)
+		return e.ErrSetBreakpoint.AddDesc("修改断点意外失败: get kube rest config")
+	}
+
+	// job task is running, check whether shell step has run, and touch breakpoint file
+	// if job task status is debug_after, only breakpoint operation can do is unset breakpoint_after, which should be done by StopDebugWorkflowTaskJobV4
+	// if job task status is prepare, setting breakpoints has a low probability of not taking effect, and the current design allows for this flaw
+	if task.Status == config.StatusRunning || task.Status == config.StatusDebugBefore || task.Status == config.StatusPrepare {
+		pods, err := getter.ListPods(jobTaskSpec.Properties.Namespace, labels.Set{"job-name": task.K8sJobName}.AsSelector(), kubeClient)
+		if err != nil {
+			logger.Errorf("set workflowTaskV4 breakpoint failed: list pods %v", err)
+			return e.ErrSetBreakpoint.AddDesc("修改断点意外失败: ListPods")
+		}
+		if len(pods) == 0 {
+			logger.Error("set workflowTaskV4 breakpoint failed: list pods num 0")
+			return e.ErrSetBreakpoint.AddDesc("修改断点意外失败: ListPods num 0")
+		}
+		pod := pods[0]
+		switch pod.Status.Phase {
+		case corev1.PodRunning:
+		default:
+			logger.Errorf("set workflowTaskV4 breakpoint failed: pod status is %s", pod.Status.Phase)
+			return e.ErrSetBreakpoint.AddDesc(fmt.Sprintf("当前任务状态 %s 无法修改断点", pod.Status.Phase))
+		}
+		exec := func(cmd string) bool {
+			opt := podexec.ExecOptions{
+				Namespace:     jobTaskSpec.Properties.Namespace,
+				PodName:       pod.Name,
+				ContainerName: pod.Spec.Containers[0].Name,
+				Command:       []string{"sh", "-c", cmd},
+			}
+			_, stderr, success, _ := podexec.KubeExec(clientSet, restConfig, opt)
+			logger.Errorf("set workflowTaskV4 breakpoint exec %s error: %s", cmd, stderr)
+			return success
+		}
+		touchOrRemove := func(set bool) string {
+			if set {
+				return "touch"
+			}
+			return "rm"
+		}
+		switch position {
+		case "before":
+			if exec(checkShellStepStart) {
+				logger.Error("set workflowTaskV4 before breakpoint failed: shell step has started")
+				return e.ErrSetBreakpoint.AddDesc("当前任务已开始运行脚本，无法修改前断点")
+			}
+			exec(fmt.Sprintf(setOrUnsetBreakpoint, touchOrRemove(set), position))
+		case "after":
+			if exec(checkShellStepDone) {
+				logger.Error("set workflowTaskV4 after breakpoint failed: shell step has been done")
+				return e.ErrSetBreakpoint.AddDesc("当前任务已运行完脚本，无法修改后断点")
+			}
+			exec(fmt.Sprintf(setOrUnsetBreakpoint, touchOrRemove(set), position))
+		}
+		// update data in memory and ack
+		switch position {
+		case "before":
+			task.BreakpointBefore = set
+			ack = c.ack
+		case "after":
+			task.BreakpointAfter = set
+			ack = c.ack
+		}
+		logger.Infof("set workflowTaskV4 breakpoint success: %s-%s %v", jobName, position, set)
+		return nil
+	}
+	logger.Errorf("set workflowTaskV4 breakpoint failed: job status is %s", task.Status)
+	return e.ErrSetBreakpoint.AddDesc("当前任务状态无法修改断点 ")
+}
+
+func (c *workflowCtl) enableWorkflowDebug() error {
+	workflowDebugLock := cache.NewRedisLockWithExpiry(WorkflowDebugLockKey(c.workflowTask.WorkflowName, c.workflowTask.TaskID), time.Second*5)
+	err := workflowDebugLock.Lock()
+	if err != nil {
+		return e.ErrStopDebugShell.AddDesc(fmt.Sprintf("failed to acquire lock when setting breakpoint, err: %s", err))
+	}
+	defer workflowDebugLock.Unlock()
+
+	if c.workflowTask.IsDebug {
+		return e.ErrStopDebugShell.AddDesc("任务已开启调试模式")
+	}
+	c.workflowTask.IsDebug = true
+	c.logger.Infof("enable workflowTaskV4 debug mode success: %s-%d", c.workflowTask.WorkflowName, c.workflowTask.TaskID)
+	c.ack()
+	return nil
+}
+
+func (c *workflowCtl) stopWorkflowDebug(jobName, position string) error {
+	workflowDebugLock := cache.NewRedisLockWithExpiry(WorkflowDebugLockKey(c.workflowTask.WorkflowName, c.workflowTask.TaskID), time.Second*5)
+	err := workflowDebugLock.Lock()
+	if err != nil {
+		return e.ErrStopDebugShell.AddDesc(fmt.Sprintf("failed to acquire lock when setting breakpoint, err: %s", err))
+	}
+	defer workflowDebugLock.Unlock()
+
+	logger := c.logger
+	var ack func()
+	defer func() {
+		if ack != nil {
+			ack()
+		}
+	}()
+
+	var task *commonmodels.JobTask
+FOR:
+	for _, stage := range c.workflowTask.Stages {
+		for _, jobTask := range stage.Jobs {
+			if jobTask.Name == jobName {
+				task = jobTask
+				break FOR
+			}
+		}
+	}
+	if task == nil {
+		logger.Error("stop workflowTaskV4 debug shell failed: not found job")
+		return e.ErrStopDebugShell.AddDesc("Job不存在")
+	}
+	jobTaskSpec := &commonmodels.JobTaskFreestyleSpec{}
+	if err := commonmodels.IToi(task.Spec, jobTaskSpec); err != nil {
+		logger.Errorf("stop workflowTaskV4 debug shell failed: IToi %v", err)
+		return e.ErrStopDebugShell.AddDesc("结束调试意外失败")
+	}
+
+	kubeClient, err := kubeclient.GetKubeClient(config.HubServerAddress(), jobTaskSpec.Properties.ClusterID)
+	if err != nil {
+		logger.Errorf("stop workflowTaskV4 debug shell failed: get kube client error: %s", err)
+		return e.ErrSetBreakpoint.AddDesc("结束调试意外失败: get kube client")
+	}
+	clientSet, err := kubeclient.GetClientset(config.HubServerAddress(), jobTaskSpec.Properties.ClusterID)
+	if err != nil {
+		logger.Errorf("stop workflowTaskV4 debug shell failed: get kube client set error: %s", err)
+		return e.ErrSetBreakpoint.AddDesc("结束调试意外失败: get kube client set")
+	}
+	restConfig, err := kubeclient.GetRESTConfig(config.HubServerAddress(), jobTaskSpec.Properties.ClusterID)
+	if err != nil {
+		logger.Errorf("stop workflowTaskV4 debug shell failed: get kube rest config error: %s", err)
+		return e.ErrSetBreakpoint.AddDesc("结束调试意外失败: get kube rest config")
+	}
+
+	pods, err := getter.ListPods(jobTaskSpec.Properties.Namespace, labels.Set{"job-name": task.K8sJobName}.AsSelector(), kubeClient)
+	if err != nil {
+		logger.Errorf("stop workflowTaskV4 debug shell failed: list pods %v", err)
+		return e.ErrStopDebugShell.AddDesc("结束调试意外失败: ListPods")
+	}
+	if len(pods) == 0 {
+		logger.Error("stop workflowTaskV4 debug shell failed: list pods num 0")
+		return e.ErrStopDebugShell.AddDesc("结束调试意外失败: ListPods num 0")
+	}
+	pod := pods[0]
+	switch pod.Status.Phase {
+	case corev1.PodRunning:
+	default:
+		logger.Errorf("stop workflowTaskV4 debug shell failed: pod status is %s", pod.Status.Phase)
+		return e.ErrStopDebugShell.AddDesc(fmt.Sprintf("Job 状态 %s 无法结束调试", pod.Status.Phase))
+	}
+	exec := func(cmd string) bool {
+		opt := podexec.ExecOptions{
+			Namespace:     jobTaskSpec.Properties.Namespace,
+			PodName:       pod.Name,
+			ContainerName: pod.Spec.Containers[0].Name,
+			Command:       []string{"sh", "-c", cmd},
+		}
+		_, stderr, success, _ := podexec.KubeExec(clientSet, restConfig, opt)
+		logger.Errorf("stop workflowTaskV4 debug shell exec %s error: %s", cmd, stderr)
+		return success
+	}
+
+	if !exec(fmt.Sprintf("ls /zadig/debug/breakpoint_%s", position)) {
+		logger.Errorf("set workflowTaskV4 %s breakpoint failed: not found file", position)
+		return e.ErrStopDebugShell.AddDesc("未找到断点文件")
+	}
+	exec(fmt.Sprintf("rm /zadig/debug/breakpoint_%s", position))
+
+	ack = c.ack
+	logger.Infof("stop workflowTaskV4 debug shell success: %s-%d", c.workflowTask.WorkflowName, c.workflowTask.TaskID)
+	return nil
 }
 
 func updateworkflowStatus(workflow *commonmodels.WorkflowTask) {
@@ -215,6 +521,68 @@ func updateworkflowStatus(workflow *commonmodels.WorkflowTask) {
 	}
 	workflow.Status = workflowStatus
 }
+
+//func UpdateWorkflowTask(workflowName string, taskID int64, logger *zap.SugaredLogger) {
+//	taskInColl, err := commonrepo.NewworkflowTaskv4Coll().Find(workflowName, taskID)
+//	if err != nil {
+//		logger.Errorf("find workflow task v4 %s failed,error: %v", workflowName, err)
+//		return
+//	}
+//	// 如果当前状态已经通过或者失败, 不处理新接受到的ACK
+//	if taskInColl.Status == config.StatusPassed || taskInColl.Status == config.StatusFailed || taskInColl.Status == config.StatusTimeout || taskInColl.Status == config.StatusReject {
+//		logger.Infof("%s:%d:%s task already done", workflowName, taskID, taskInColl.Status)
+//		return
+//	}
+//	// 如果当前状态已经是取消状态, 一般为用户取消了任务, 此时任务在短暂时间内会继续运行一段时间,
+//	if taskInColl.Status == config.StatusCancelled {
+//		// Task终止状态可能为Pass, Fail, Cancel, Timeout
+//		// backend 会继续接受到ACK, 在这种情况下, 终止状态之外的ACK都无需处理，避免出现取消之后又被重置成运行态
+//		if taskInColl.Finished() {
+//			logger.Infof("%s:%d task has been cancelled, ACK dropped", workflowName, taskID)
+//			return
+//		}
+//		//if c.workflowTask.Status != config.StatusFailed && c.workflowTask.Status != config.StatusPassed && c.workflowTask.Status != config.StatusCancelled && c.workflowTask.Status != config.StatusTimeout {
+//		//	logger.Infof("%s:%d task has been cancelled, ACK dropped", workflowName, taskID)
+//		//	return
+//		//}
+//	}
+//	if success := UpdateQueue(c.workflowTask); !success {
+//		logger.Errorf("%s:%d update t status error", workflowName, taskID)
+//	}
+//	// TODO update workflow task
+//	if err := commonrepo.NewworkflowTaskv4Coll().Update(c.workflowTask.ID.Hex(), c.workflowTask); err != nil {
+//		logger.Errorf("update workflow task v4 failed,error: %v", err)
+//	}
+//
+//	if c.workflowTask.Status == config.StatusPassed || c.workflowTask.Status == config.StatusFailed || c.workflowTask.Status == config.StatusTimeout || c.workflowTask.Status == config.StatusCancelled || c.workflowTask.Status == config.StatusReject {
+//		logger.Infof("%s:%d:%v task done", c.workflowTask.WorkflowName, c.workflowTask.TaskID, c.workflowTask.Status)
+//		if err := instantmessage.NewWeChatClient().SendWorkflowTaskNotifications(c.workflowTask); err != nil {
+//			logger.Errorf("send workflow task notification failed, error: %v", err)
+//		}
+//		q := ConvertTaskToQueue(c.workflowTask)
+//		if err := Remove(q); err != nil {
+//			logger.Errorf("remove queue task: %s:%d error: %v", workflowName, taskID, err)
+//		}
+//		result, err := commonrepo.NewStrategyColl().GetByTarget(commonmodels.WorkflowTaskRetention)
+//		if err != nil {
+//			logger.Errorf("get workflow task retention strategy error: %s", err)
+//			result = commonmodels.DefaultWorkflowTaskRetention
+//		}
+//		if err = commonrepo.NewworkflowTaskv4Coll().ArchiveHistoryWorkflowTask(workflowName, result.Retention.MaxItems, result.Retention.MaxDays); err != nil {
+//			logger.Errorf("ArchiveHistoryWorkflowTask error: %v", err)
+//		}
+//		// Updating the comment in the git repository, this will not cause the function to return error if this function call fails
+//		if err := scmnotify.NewService().UpdateWebhookCommentForWorkflowV4(c.workflowTask, logger); err != nil {
+//			log.Warnf("Failed to update comment for custom workflow %s, taskID: %d the error is: %s", workflowName, taskID, err)
+//		}
+//		if err := scmnotify.NewService().CompleteGitCheckForWorkflowV4(c.workflowTask.WorkflowArgs, taskID, c.workflowTask.Status, logger); err != nil {
+//			log.Warnf("Failed to update github check status for custom workflow %s, taskID: %d the error is: %s", workflowName, taskID, err)
+//		}
+//		if err := workflowstat.UpdateWorkflowStat(c.workflowTask.WorkflowName, string(config.WorkflowTypeV4), string(c.workflowTask.Status), c.workflowTask.ProjectName, c.workflowTask.EndTime-c.workflowTask.StartTime, c.workflowTask.IsRestart); err != nil {
+//			log.Warnf("Failed to update workflow stat for custom workflow %s, taskID: %d the error is: %s", workflowName, taskID, err)
+//		}
+//	}
+//}
 
 func (c *workflowCtl) updateWorkflowTask() {
 	taskInColl, err := commonrepo.NewworkflowTaskv4Coll().Find(c.workflowTask.WorkflowName, c.workflowTask.TaskID)
@@ -316,6 +684,7 @@ func (c *workflowCtl) CleanShareStorage() {
 	}
 
 	cache.NewRedisCache(config2.RedisCommonCacheTokenDB()).Delete(c.prefix + "-cluster")
+	cache.NewRedisCache(config2.RedisCommonCacheTokenDB()).Delete(c.prefix + "-debug")
 	cache.NewRedisCache(config2.RedisCommonCacheTokenDB()).Delete(c.prefix)
 }
