+++ pkg/tool/errors/http_errors.go
@@ -322,6 +322,8 @@ var (
 
 	// ErrGetDebugShell
 	ErrGetDebugShell = NewHTTPError(6172, "获取调试 Shell 失败")
+
+	ErrEnableDebug = NewHTTPError(6173, "开启工作流任务调试失败")
 	//-----------------------------------------------------------------------------------------------
 	// Keystore APIs Range: 6180 - 6189
 	//-----------------------------------------------------------------------------------------------
